+++ rust/crates/turbo-surf-napi/src/lib.rs
@@ -25,7 +25,7 @@ use view::{Field, FieldType, QueryType, TextMode};
 
 #[napi]
 pub fn version() -> String {
-    "0.2.6".to_string()
+    "0.2.7".to_string()
 }
 
 fn to_json_string<T: serde::Serialize>(v: &T) -> String {